+++ src/parser/parse_tree.h
@@ -506,7 +506,7 @@
 #define PT_IS_NUMBERING_AFTER_EXECUTION(op) \
         ( ((op) == PT_INST_NUM || \
            (op) == PT_ROWNUM || \
-           (op) == PT_GROUPBY_NUM || \
+           (int)(op) == (int)PT_GROUPBY_NUM || \
            (op) == PT_ORDERBY_NUM) ? true : false )
 
 #define PT_IS_SERIAL(op) \
